move likedby and repostedby to new file structure
